Fix ILogger resolution for ACTNARs
This fixes almost all "UnknownLogMessageSource" warnings printed. We ask
the container to construct concrete types in many places, and since
these registrations are provided on-the-fly, they are not sent through
the normal LoggerModule. This commit adds the same resolution step for
those components that are registered with the
SelfBindingRegistrationSource, which removes the problem for those
registrations.