Add missing logger argument to ExecutionContextManager.
It was causing just a bit under a million exceptions.

It seems like the Microsoft loggers throw exceptions if the number of
holes in the formatted log template does not match the number of
arguments passed. These errors are now just catched, which means they
silently dissapear. Not sure what to do with that, but we should at
in the future figure out how to get a handle on those kinds of errors.